Update README.md
I'm sorry but I added my project idea to the wrong place in Github, where I described my project earlier in the semester. I apologize that this entry is late, please let me know if there are other things that need to be added to my code, though I ran it in Colaboratory and it runs fine.